move : (enum...) 파일 이동
- 엔티티 상위에 위치로 이동
